images/imposm/requirements.txt
mapcoloniesjsonlogger==0.0.14
psycopg2==2.8.6
